Completing basic JUnit testing on database
After the creation of tables, mock data is inserted
and retrieved. The retrieved query values are then
compared with the original mock data to ensure that
values and data type are preserved.

Additionally, scene, message and choice records cannot
be inserted without an existing record in their respective
dependent tables, so referential integrity is also tested
at a basic level.

A slight change is also made to the choice table schema
to reduce a redundancy in timestamping. Only the date of
choice selection must be recorded by the client. This field
is created with a null value initially.